web/requirements.txt
gunicorn
flask
flask_sqlalchemy
flask_wtf 
flask_login
psycopg2
jinja2
wtforms
pytesseract
Pillow
